requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
joblib
matplotlib
seaborn
numpy
pymongo
pydantic
pydantic_settings
python-dotenv
ipykernel